refactor: modularize world setup
