+++ main.js
@@ -8,166 +8,26 @@ import { initMinimap, updateMinimap } from './minimap.js';
 import { initMainMenu } from './src/ui/mainMenu.js';
 import { loadSettings, saveSettings } from './src/state/persistence.js';
 import { gsap } from 'https://cdn.jsdelivr.net/npm/gsap@3.13.0/index.js';
+import {
+  TILE, WORLD_W, WORLD_H, DPR, DAY_LENGTH_SEC,
+  WEATHER, WEATHER_NAMES, BIOME, COLORS,
+  plant, terrain, soilMoisture,
+  crowdSmall, crowdMedium, crowdLarge, pollinatorBoost,
+  animals, SMALL_LIMIT, LARGE_LIMIT, CROWD_THRESH, CROWD_DECAY,
+  GRID_SIZE, GRID_W, GRID_H, cellIndex, grid,
+  idx, clamp, speciesConfig
+} from './src/worldConfig.js';
+import { generateTerrain, generateSoilMoisture, seededRandom } from './src/terrain.js';
+import {
+  getWorldTime, setWorldTime,
+  getSimTime, setSimTime,
+  getWeatherState, setWeatherState,
+  getWeatherTimer, setWeatherTimer,
+  isNight, daylightFactor, advanceWeather
+} from './src/weather.js';
+import { runSelfTests } from './src/selfTests.js';
+
 let state;
-// ==============================================================
-//                    PARÁMETROS DEL MUNDO
-// ==============================================================
-const TILE = 10;             // Tamaño del tile en píxeles en el lienzo
-const WORLD_W = 100;         // Número de tiles horizontales
-const WORLD_H = 60;          // Número de tiles verticales
-const DPR = Math.max(1, Math.min(2, window.devicePixelRatio || 1)); // Densidad de píxeles (cap 2x para rendimiento)
-
-// Ciclo día/noche (24h simuladas en X segundos)
-const DAY_LENGTH_SEC = 120;  // Aumenta/disminuye para acelerar o frenar el ciclo
-
-// Estados de clima
-const WEATHER = { CLEAR:0, RAIN:1, DROUGHT:2 };
-const WEATHER_NAMES = ['Despejado','Lluvia','Sequía'];
-
-// Tipos de terreno por tile (mapa discreto)
-const BIOME = { WATER:0, DIRT:1, GRASS:2, BARRIER:3 };
-
-// Paleta de colores para el render
-const COLORS = {
-  WATER: '#1e40af',
-  SHORE: '#2563eb',
-  DIRT:  '#6b4226',
-  GRASS0:'#14532d',
-  GRASS1:'#166534',
-  GRASS2:'#22c55e',
-  BARRIER:'#64748b'
-};
-
-// Campos escalares del mundo:
-// - plant: densidad de vegetación 0..1 por tile
-// - terrain: tipo de bioma por tile
-// - soilMoisture: humedad del suelo 0..1 por tile
-const plant = new Float32Array(WORLD_W * WORLD_H);
-const terrain = new Uint8Array(WORLD_W * WORLD_H);
-const soilMoisture = new Float32Array(WORLD_W * WORLD_H);
-
-// Per-tile crowding counters grouped by size class
-const crowdSmall = new Float32Array(WORLD_W * WORLD_H);
-const crowdMedium = new Float32Array(WORLD_W * WORLD_H);
-const crowdLarge = new Float32Array(WORLD_W * WORLD_H);
-const pollinatorBoost = new Uint8Array(WORLD_W * WORLD_H);
-
-// Size thresholds and crowding control parameters
-const SMALL_LIMIT = 0.30;      // radius < SMALL_LIMIT => small
-const LARGE_LIMIT = 0.36;      // radius >= LARGE_LIMIT => large
-const CROWD_THRESH = { small:6, medium:4, large:2 };
-const CROWD_DECAY = 0.6;       // decay factor applied each step
-
-// Configuración de especies y poblaciones iniciales
-// Cada especie define parámetros básicos y dieta de recursos o presas
-const speciesConfig = {
-  HERB: {
-    start: 28,
-    baseSpeed: 0.55,
-    hungerRate: 0.015,           // Energía perdida por segundo
-    vision: 6,
-    preyEnergy: 0.9,
-    fecundity: 1.6,
-    size: 'medium',
-    lifespan: 200,
-    reproThreshold: 1.6,         // Energía mínima para reproducirse
-    reproCost: 0.8,
-    reproCooldown: 5,
-    nightSpeedMul: 0.9,
-    nightVisionMul: 0.9,
-    maxEnergy: 2.5,
-    radius: 0.32,
-    initEnergy: 1.0,
-    addEnergy: 1.0,
-    offspringEnergy: 0.8,
-    wanderFactor: 2.0,
-    diet: { PLANT: { energy: 0.18, plantDelta: 0.22 } }
-  },
-  CARN: {
-    start: 10,
-    baseSpeed: 0.7,
-    hungerRate: 0.02,
-    vision: 9,
-    preyEnergy: 1.2,
-    fecundity: 2.2,
-    size: 'medium',
-    lifespan: 220,
-    reproThreshold: 2.2,
-    reproCost: 1.1,
-    reproCooldown: 7,
-    nightSpeedMul: 1.1,
-    nightVisionMul: 1.1,
-    radius: 0.36,
-    initEnergy: 1.4,
-    addEnergy: 1.2,
-    offspringEnergy: 1.0,
-    wanderFactor: 1.3,
-    diet: { HERB: true }
-  },
-  RODENT: {
-    start: 12,
-    baseSpeed: 0.5,
-    hungerRate: 0.02,
-    vision: 5,
-    preyEnergy: 0.8,
-    fecundity: 1.2,
-    size: 'small',
-    lifespan: 180,
-    reproThreshold: 1.2,
-    reproCost: 0.6,
-    reproCooldown: 5,
-    nightSpeedMul: 0.9,
-    nightVisionMul: 0.9,
-    radius: 0.3,
-    initEnergy: 0.9,
-    offspringEnergy: 0.7,
-    wanderFactor: 2.0,
-    diet: { PLANT: { energy: 0.15, plantDelta: 0.2 } }
-  },
-  WOLF: {
-    start: 4,
-    baseSpeed: 0.8,
-    hungerRate: 0.025,
-    vision: 10,
-    preyEnergy: 1.5,
-    fecundity: 2.4,
-    size: 'large',
-    lifespan: 240,
-    reproThreshold: 2.4,
-    reproCost: 1.2,
-    reproCooldown: 7,
-    nightSpeedMul: 1.1,
-    nightVisionMul: 1.1,
-    radius: 0.38,
-    initEnergy: 1.5,
-    offspringEnergy: 1.1,
-    wanderFactor: 1.3,
-    diet: { HERB: true, RODENT: true }
-  },
-  POLLINATOR: {
-    start: 16,
-    baseSpeed: 0.5,
-    hungerRate: 0.01,
-    vision: 4,
-    preyEnergy: 0.5,
-    fecundity: 1.0,
-    size: 'small',
-    lifespan: 160,
-    reproThreshold: 1.0,
-    reproCost: 0.5,
-    reproCooldown: 5,
-    nightSpeedMul: 1.0,
-    nightVisionMul: 1.0,
-    radius: 0.28,
-    initEnergy: 0.8,
-    offspringEnergy: 0.6,
-    wanderFactor: 2.0,
-    diet: { PLANT: { energy: 0.05, plantDelta: 0.005 } }
-  }
-};
-
-// Array dinámico de individuos; cada uno es un objeto con estado y genes
-const animals = [];
 
 const speciesList = Object.keys(speciesConfig);
 const spawnEnabled = Object.fromEntries(speciesList.map(s=>[s,true]));
@@ -183,17 +43,6 @@ if (settings.hiddenSpecies) Object.assign(hiddenSpecies, settings.hiddenSpecies)
 if (settings.spawnRate) Object.assign(spawnRate, settings.spawnRate);
 if (settings.reproThresholdMul) Object.assign(reproThresholdMul, settings.reproThresholdMul);
 if (settings.mortalityMul) Object.assign(mortalityMul, settings.mortalityMul);
-
-// Spatial grid for neighborhood queries (optimizes nearest searches)
-const GRID_SIZE = 10; // tiles per cell
-const GRID_W = Math.ceil(WORLD_W / GRID_SIZE);
-const GRID_H = Math.ceil(WORLD_H / GRID_SIZE);
-const cellIndex = (gx, gy)=> gy*GRID_W + gx;
-const grid = Array.from({length: GRID_W*GRID_H}, () => []);
-
-// Utilidades cortas para índices y límites
-const idx = (x,y)=> y*WORLD_W + x;                   // Indexa (x,y) en arreglos lineales
-const clamp = (v, a, b)=> Math.max(a, Math.min(b, v)); // Limita v al rango [a,b]
 
 // ==============================================================
 //                         CANVAS SETUP
@@ -229,115 +78,6 @@ function resizeCanvas() {
 }
 resizeCanvas();
 window.addEventListener('resize', resizeCanvas);
-
-// ==============================================================
-//                    RUIDO / MAPA DE ALTURA
-// ==============================================================
-function seededRandom(seed) {
-  // PRNG congruencial lineal reproducible
-  let s = seed >>> 0;
-  return function() {
-    s = (s * 1664525 + 1013904223) >>> 0;
-    return (s & 0xfffffff) / 0xfffffff; // Devuelve [0,1)
-  };
-}
-
-function valueNoise2D(width, height, gridStep, rng){
-  // Genera ruido 2D por interpolación bilineal de una grilla de valores aleatorios
-  const gw = Math.ceil(width / gridStep)+2;
-  const gh = Math.ceil(height/ gridStep)+2;
-  const grid = new Float32Array(gw*gh);
-  for(let gy=0; gy<gh; gy++){
-    for(let gx=0; gx<gw; gx++){
-      grid[gy*gw+gx] = rng();
-    }
-  }
-  const out = new Float32Array(width*height);
-  for(let y=0; y<height; y++){
-    const gy = Math.floor(y / gridStep);
-    const fy = (y / gridStep) - gy;           // Fracción vertical entre celdas
-    for(let x=0; x<width; x++){
-      const gx = Math.floor(x / gridStep);
-      const fx = (x / gridStep) - gx;         // Fracción horizontal
-      // Cuatro esquinas de la celda
-      const a = grid[gy*gw+gx];
-      const b = grid[gy*gw+gx+1];
-      const c = grid[(gy+1)*gw+gx];
-      const d = grid[(gy+1)*gw+gx+1];
-      // Suavizado cúbico (curva S) para evitar artefactos
-      const sx = fx*fx*(3-2*fx);
-      const sy = fy*fy*(3-2*fy);
-      const i1 = a + (b-a)*sx;                // Interpola horizontal arriba
-      const i2 = c + (d-c)*sx;                // Interpola horizontal abajo
-      out[y*width+x] = i1 + (i2 - i1)*sy;     // Interpola vertical
-    }
-  }
-  return out;
-}
-
-// Construye terreno base a partir de 2 octavas de ruido
-function generateTerrain() {
-  const rng = seededRandom(1337);
-  const n1 = valueNoise2D(WORLD_W, WORLD_H, 8, rng);
-  const n2 = valueNoise2D(WORLD_W, WORLD_H, 4, rng);
-
-  for(let y=0; y<WORLD_H; y++){
-    for(let x=0; x<WORLD_W; x++){
-      // Altura sintética 0..1; umbrales definen agua/orilla/pasto
-      const h = n1[idx(x,y)]*0.7 + n2[idx(x,y)]*0.3;
-      if (h < 0.35) {
-        terrain[idx(x,y)] = BIOME.WATER;  // Agua: no crecen plantas
-        plant[idx(x,y)] = 0.0;
-      } else if (h < 0.45) {
-        terrain[idx(x,y)] = BIOME.DIRT;   // Orilla/playa
-        plant[idx(x,y)] = 0.0;
-      } else {
-        terrain[idx(x,y)] = BIOME.GRASS;  // Zona fértil
-        plant[idx(x,y)] = Math.max(0, (h-0.45)*1.4);
-      }
-    }
-  }
-}
-
-function generateSoilMoisture() {
-  const rng = seededRandom(4242);
-  const n1 = valueNoise2D(WORLD_W, WORLD_H, 8, rng);
-  const n2 = valueNoise2D(WORLD_W, WORLD_H, 4, rng);
-  for (let i = 0; i < soilMoisture.length; i++) {
-    soilMoisture[i] = clamp(n1[i] * 0.7 + n2[i] * 0.3, 0, 1);
-  }
-}
-
-// ==============================================================
-//                    TIEMPO DEL MUNDO Y CLIMA
-// ==============================================================
-let worldTime = 0; // Segundos simulados acumulados
-let simTime = 0;   // Segundos de simulación para UI
-let weatherState = WEATHER.CLEAR;
-let weatherTimer = 0; // Cuenta atrás del estado de clima actual
-
-function isNight(){
-  // Noche cuando el tiempo normalizado está cerca de 0 o 1 (cuartos del día)
-  const t = (worldTime % DAY_LENGTH_SEC) / DAY_LENGTH_SEC; // 0..1
-  return (t < 0.25) || (t > 0.75);
-}
-function daylightFactor(){
-  // Factor de luz ambiental (0..1) como seno; máximo a mediodía
-  const t = (worldTime % DAY_LENGTH_SEC) / DAY_LENGTH_SEC;
-  return 0.35 + 0.65 * Math.max(0, Math.sin(Math.PI * (t)));
-}
-function advanceWeather(dt){
-  // Manejo semi-Markoviano simple: cuando expira el temporizador, elige un nuevo clima
-  weatherTimer -= dt;
-  if (weatherTimer <= 0){
-    const r = Math.random();
-    weatherState = (r < 0.55) ? WEATHER.CLEAR : (r < 0.85 ? WEATHER.RAIN : WEATHER.DROUGHT);
-    // Duración base según estado con jitter aleatorio
-    const base = weatherState===WEATHER.RAIN ? 18 : (weatherState===WEATHER.DROUGHT ? 20 : 24);
-    const jitter = (Math.random()*0.6+0.7); // 0.7..1.3
-    weatherTimer = base * jitter;
-  }
-}
 
 // ==============================================================
 //                       POBLACIÓN / GENÉTICA
@@ -424,15 +164,15 @@ function growPlants(){
 
       // Actualiza humedad del suelo
       let m = soilMoisture[id];
-      if (weatherState===WEATHER.RAIN) m += 0.04;
-      if (weatherState===WEATHER.DROUGHT) m -= 0.03;
+      if (state.weatherState===WEATHER.RAIN) m += 0.04;
+      if (state.weatherState===WEATHER.DROUGHT) m -= 0.03;
       m -= 0.005; // evaporación base
       soilMoisture[id] = clamp(m, 0, 1);
 
       if (terrain[id] !== BIOME.GRASS) continue; // Solo crece en pasto
       let inc = 0.02;                            // Ritmo base
-      if (weatherState===WEATHER.RAIN) inc += 0.03;     // Lluvia acelera
-      if (weatherState===WEATHER.DROUGHT) inc -= 0.02;  // Sequía frena
+      if (state.weatherState===WEATHER.RAIN) inc += 0.03;     // Lluvia acelera
+      if (state.weatherState===WEATHER.DROUGHT) inc -= 0.02;  // Sequía frena
       if (night) inc -= 0.005;                          // Noche penaliza un poco
       // Detecta agua en vecindad 8-conexa (beneficia crecimiento)
       let nearWater = false;
@@ -465,7 +205,7 @@ function growPlants(){
         if (d2 < fire.r*fire.r) inc -= 0.15; // Quemado fuerte
       }
       plant[id] = clamp(plant[id] + inc, 0, 1);
-      if (weatherState===WEATHER.DROUGHT) plant[id] = Math.max(0, plant[id] - 0.01);
+      if (state.weatherState===WEATHER.DROUGHT) plant[id] = Math.max(0, plant[id] - 0.01);
     }
   }
 }
@@ -622,14 +362,14 @@ let flashTimer = 0; // Cuenta atrás del flash de meteorito
 
 // Objeto de estado compartido entre módulos
 state = {
-  get simTime(){ return simTime; },
-  set simTime(v){ simTime = v; },
-  get worldTime(){ return worldTime; },
-  set worldTime(v){ worldTime = v; },
-  get weatherTimer(){ return weatherTimer; },
-  set weatherTimer(v){ weatherTimer = v; },
-  get weatherState(){ return weatherState; },
-  set weatherState(v){ weatherState = v; },
+  get simTime(){ return getSimTime(); },
+  set simTime(v){ setSimTime(v); },
+  get worldTime(){ return getWorldTime(); },
+  set worldTime(v){ setWorldTime(v); },
+  get weatherTimer(){ return getWeatherTimer(); },
+  set weatherTimer(v){ setWeatherTimer(v); },
+  get weatherState(){ return getWeatherState(); },
+  set weatherState(v){ setWeatherState(v); },
   get accPlant(){ return accPlant; },
   set accPlant(v){ accPlant = v; },
   get flashTimer(){ return flashTimer; },
@@ -710,7 +450,7 @@ function loop(now){
     fps = Math.round(frames / fpsTime);
     frames = 0; fpsTime = 0;
     if ($fps) $fps.textContent = `FPS: ${fps}`;
-    if ($tick) $tick.textContent = `t: ${simTime.toFixed(1)}s`;
+    if ($tick) $tick.textContent = `t: ${state.simTime.toFixed(1)}s`;
     updateHUD(state);
   }
 
@@ -736,75 +476,4 @@ gsap.to('#evtFire', { rotation:-10, yoyo:true, repeat:-1, duration:0.6, transfor
 gsap.to('#evtMeteor', { y:-4, yoyo:true, repeat:-1, duration:0.8 });
 gsap.to('#evtPlagueH, #evtPlagueC', { scale:1.1, yoyo:true, repeat:-1, duration:1 });
 
-// ==============================================================
-//                       SELF-TESTS (sanidad)
-// ==============================================================
-function runSelfTests(){
-  const tests = [];
-  const add = (name, pass, info='') => tests.push({ name, pass, info });
-
-  // Tamaños de buffers del mundo
-  add('terrain tamaño', terrain.length === WORLD_W*WORLD_H);
-  add('soilMoisture tamaño', soilMoisture.length === WORLD_W*WORLD_H);
-  // Asegura rango [0,1] en plantas
-  add('plant [0,1]', (()=>{ for (let i=0;i<plant.length;i++){ const v=plant[i]; if(!(v>=0 && v<=1)) return false; } return true; })());
-  add('soilMoisture [0,1]', (()=>{ for (let i=0;i<soilMoisture.length;i++){ const v=soilMoisture[i]; if(!(v>=0 && v<=1)) return false; } return true; })());
-
-  // Poblaciones iniciales
-  const h0 = animals.filter(a=>a.sp==='HERB').length;
-  const c0 = animals.filter(a=>a.sp==='CARN').length;
-  const r0 = animals.filter(a=>a.sp==='RODENT').length;
-  const w0 = animals.filter(a=>a.sp==='WOLF').length;
-  const p0 = animals.filter(a=>a.sp==='POLLINATOR').length;
-  add('HERB_START', h0 === speciesConfig.HERB.start, `h0=${h0}`);
-  add('CARN_START', c0 === speciesConfig.CARN.start, `c0=${c0}`);
-  add('RODENT_START', r0 === speciesConfig.RODENT.start, `r0=${r0}`);
-  add('WOLF_START', w0 === speciesConfig.WOLF.start, `w0=${w0}`);
-  add('POLL_START', p0 === speciesConfig.POLLINATOR.start, `p0=${p0}`);
-
-  // UI básica presente
-  add('Toolbar presente', !!document.getElementById('toolbar'));
-
-  // Pruebas de herramientas (sobre un tile de GRASS real si existe)
-  const findGrassTile = ()=>{
-    for(let yy=0; yy<WORLD_H; yy++){
-      for(let xx=0; xx<WORLD_W; xx++){
-        if (terrain[idx(xx,yy)]===BIOME.GRASS) return {x:xx,y:yy};
-      }
-    }
-    return {x:5,y:5}; // Fallback (poco probable no hallar GRASS)
-  };
-  const {x:gx,y:gy} = findGrassTile();
-  const idG = idx(gx,gy); const oldPlant=plant[idG];
-  applyActionAt(state,gx,gy,TOOL.FOOD); add('Comida aumenta planta', plant[idG] >= oldPlant);
-
-  // Verifica FIX de RNG: defaultGenes sin rng debe funcionar y estar en rango
-  let okGenes = true; let g;
-  try { g = defaultGenes('HERB'); } catch(e){ okGenes = false; }
-  add('defaultGenes() sin rng no falla', okGenes);
-  if (okGenes){
-    add('speedMul rango', g.speedMul>=0.75 && g.speedMul<=1.35, `v=${g.speedMul.toFixed(3)}`);
-    add('metabolismMul rango', g.metabolismMul>=0.7 && g.metabolismMul<=1.5, `v=${g.metabolismMul.toFixed(3)}`);
-    add('visionMul rango', g.visionMul>=0.7 && g.visionMul<=1.6, `v=${g.visionMul.toFixed(3)}`);
-    add('lifespan rango', g.lifespan>=160 && g.lifespan<=240, `v=${g.lifespan}`);
-  }
-
-  // Añadir animal mediante herramienta debe incrementar población
-  const before = animals.length;
-  applyActionAt(state,gx,gy,TOOL.ADD_HERB);
-  add('ADD_HERB incrementa población', animals.length === before+1, `before=${before} now=${animals.length}`);
-
-  // Cambios de terreno por herramientas
-  applyActionAt(state,gx,gy,TOOL.WATER);   add('Agua cambia a WATER', terrain[idG] === BIOME.WATER);
-  applyActionAt(state,gx,gy,TOOL.BARRIER); add('Barrera cambia a BARRIER', terrain[idG] === BIOME.BARRIER);
-
-  // Reporte visual breve
-  const passed = tests.filter(t=>t.pass).length;
-  const el = document.createElement('div'); el.className = 'tests'; el.innerHTML = `<b>Self-tests:</b> ${passed}/${tests.length} OK`;
-  document.body.appendChild(el);
-  console.group('%cSelf-tests Fase 4','color:#60a5fa');
-  tests.forEach(t=>console[t.pass? 'log': 'error'](`${t.pass?'✔':'✖'} ${t.name}${t.info? ' — '+t.info:''}`));
-  console.groupEnd();
-}
-runSelfTests();
-
+runSelfTests(state, applyActionAt, TOOL, defaultGenes);
